Match kconfig options to a kernel's architecture
The 'necessary_kconfig_options' dictionary in pmb/config/__init__.py
now has the different architectures (space seperated) as the keys and
the dictionary, which matches kernel config options and their
expected value, as its value.
For that purpose, the 'check' function in pmb/parse/kconfig.py was
modified, so that it takes the first architecture from the APKBUILD
and uses it to find the needed kernel config options.